Use the /menu endpoint when available
Before this patch the pages to check was static. This is problematic
because plugins can add pages dynamically. The /menu endpoint returns a
JSON representation of the menu available for that user.

This change retrieves the endpoint. When it returns 404, it falls back
to the static list. On success the dynamic menu is used.