Allow passing in a subset of query parameters
Katello adds some query parameters to the actual URL. This is fine and
doesn't matter. However, if a menu item does have some query parameter
then it must also be present in the actual URL. This is achieved by
parsing the query string, reducing it to the set present in the expected
URL and then comparing.