Set --systemd always for podman client containers
The default for the --systemd option in podman is to use the default
CMD for a container to detect whether systemd should be set to true.
From testing this appears to be buggy and not to always work. By setting
--systemd to always we force podman to run with systemd working
inside the containers and avoid any potential issues with automatic
detection.