Add a plugin marker that autodetects plugins
This uses the /api/v2/plugins endpoint to list all plugins. Then
individual tests can mark a required plugin to be required. The test is
skipped when the plugin isn't present.

The loading of plugins is lazy because when pytest_configure is called,
other plugins may not have been initialized. It's also not guaranteed
that the plugins are actually called.

It doesn't implement a CLI option to fail if certain expected plugins
aren't present.